moved more from index to log
