+++ functions/config.py
@@ -4,6 +4,8 @@ defaultPrefix = "!"
 
 description = "Hello, my name is Friday, I am a chatbot built with TensorFlow and Keras, meaning I enjoy conversations. I also have a few commands for fun and moderating your servers!"
 
+reloadable_bot = "cogs.log"
+
 intents = discord.Intents(
     guilds=True,
     voice_states=True,
